# 后端/HuiXiangLife/docs/apifox/huixianglife-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "惠享生活接口文档",
    "description": "面向 Apifox 导入的全项目规划版接口文档。按 admin、user、notify 三类入口划分，接口名使用中文，便于你后续直接对齐前后端开发。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/admin/order/page": {
      "get": {
        "tags": [
          "Admin-订单管理"
        ],
        "summary": "分页查询订单",
        "operationId": "adminOrderPage",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "orderNo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "merchantId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "productId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "beginCreateTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endCreateTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "beginPayTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endPayTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOrderDetailVO"
                }
              }
            }
          }
        }
      }
    },
    "/admin/order/{id}": {
      "get": {
        "tags": [
          "Admin-订单管理"
        ],
        "summary": "查询订单详情",
        "operationId": "adminOrderDetail",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailVO"
                }
              }
            }
          }
        }
      }
    },
    "/admin/order/{id}/cancel": {
      "put": {
        "tags": [
          "Admin-订单管理"
        ],
        "summary": "后台取消订单",
        "operationId": "adminOrderCancel",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
    "/admin/order/{id}/refund": {
      "post": {
        "tags": [
          "Admin-订单管理"
        ],
        "summary": "后台发起退款",
        "operationId": "adminOrderRefund",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminOrderRefundDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "退款申请成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/auth/login": {
      "post": {
        "tags": [
          "Admin-认证"
        ],
        "summary": "管理端登录",
        "operationId": "adminLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/admin/auth/logout": {
      "post": {
        "tags": [
          "Admin-认证"
        ],
        "summary": "管理端退出登录",
        "operationId": "adminLogout",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "退出成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
    "/admin/auth/me": {
      "get": {
        "tags": [
          "Admin-认证"
        ],
        "summary": "获取当前登录管理员信息",
        "operationId": "adminMe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoVO"
                }
              }
            }
          }
        }
      }
    },
    "/user/auth/register": {
      "post": {
        "tags": [
          "User-认证"
        ],
        "summary": "用户注册",
        "operationId": "userRegister",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/user/auth/login": {
      "post": {
        "tags": [
          "User-认证"
        ],
        "summary": "用户登录",
        "operationId": "userLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/user/auth/logout": {
      "post": {
        "tags": [
          "User-认证"
        ],
        "summary": "用户退出登录",
        "operationId": "userLogout",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "退出成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
    "/user/auth/me": {
      "get": {
        "tags": [
          "User-认证"
        ],
        "summary": "获取当前登录用户信息",
        "operationId": "userMe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/order/page": {
      "get": {
        "tags": [
          "User-订单"
        ],
        "summary": "分页查询我的订单",
        "operationId": "userOrderPage",
        "description": "当前登录用户从 Bearer Token 中解析，无需传 userId。",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": 0
          },
          {
            "name": "orderNo",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "HX202606170001"
          },
          {
            "name": "merchantId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1001
          },
          {
            "name": "productId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 2001
          },
          {
            "name": "beginCreateTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2026-06-17T00:00:00"
          },
          {
            "name": "endCreateTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2026-06-17T23:59:59"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOrderDetailVO"
                }
              }
            }
          }
        }
      }
    },
    "/user/order/{id}": {
      "get": {
        "tags": [
          "User-订单"
        ],
        "summary": "查询订单详情",
        "operationId": "userOrderDetail",
        "description": "只能查询当前登录用户自己的订单。",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 5001
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailVO"
                }
              }
            }
          }
        }
      }
    },
    "/user/order/{id}/cancel": {
      "post": {
        "tags": [
          "User-订单"
        ],
        "summary": "取消订单",
        "operationId": "userOrderCancel",
        "description": "只能取消当前登录用户自己的待支付订单。",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 5001
          }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "请求参数错误"
          },
          "traceId": {
            "type": "string",
            "example": "c0a80101-20260611-0001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "BooleanResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        }
      },
      "IdResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1943278456123456789
          },
          "message": {
            "type": "string",
            "example": "创建成功"
          }
        }
      },
...
      "PageMeta": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "example": 1
          },
          "size": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "example": 125
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "example": 13
          }
        }
      },
      "UserLoginDTO": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "UserRegisterDTO": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "nickname": {
            "type": "string",
            "example": "测试用户"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.png"
          }
        },
        "required": [
          "phone",
          "password",
          "nickname"
        ]
      },
...
      "AdminOrderRefundDTO": {
        "type": "object",
        "properties": {
          "refundAmount": {
            "type": "number",
            "format": "double",
            "example": 88
          },
          "reason": {
            "type": "string",
            "example": "商户无法履约"
          }
        },
        "required": [
          "refundAmount",
          "reason"
        ]
      },
...
      "LoginVO": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiJ9.demo-token"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expireIn": {
            "type": "integer",
            "example": 7200
          },
          "userInfo": {
            "$ref": "#/components/schemas/UserInfoVO"
          }
        }
      },
...
      "UserInfoVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 3001
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "nickname": {
            "type": "string",
            "example": "测试用户"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.png"
          },
          "role": {
            "type": "string",
            "example": "USER"
          },
          "status": {
            "type": "integer",
            "example": 1
          },
          "lastLoginTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-11T18:20:00"
          }
        }
      },
...
      "OrderDetailVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 5001
          },
          "orderNo": {
            "type": "string",
            "example": "HX202606110001"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 3001
          },
          "userNickname": {
            "type": "string",
            "example": "测试用户"
          },
          "merchantId": {
            "type": "integer",
            "format": "int64",
            "example": 1001
          },
          "merchantName": {
            "type": "string",
            "example": "川味小馆"
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "example": 2001
          },
          "productName": {
            "type": "string",
            "example": "双人餐套餐"
          },
          "couponId": {
            "type": "integer",
            "format": "int64",
            "example": 4001
          },
          "totalAmount": {
            "type": "number",
            "format": "double",
            "example": 176
          },
          "discountAmount": {
            "type": "number",
            "format": "double",
            "example": 20
          },
          "payAmount": {
            "type": "number",
            "format": "double",
            "example": 156
          },
          "status": {
            "type": "integer",
            "example": 1,
            "description": "0待支付 1已支付 2已完成 3已取消 4已退款"
          },
          "remark": {
            "type": "string",
            "example": "不要香菜"
          },
          "payTime": {
            "type": "string",
            "format": "date-time"
          },
          "cancelTime": {
            "type": "string",
            "format": "date-time"
          },
          "finishTime": {
            "type": "string",
            "format": "date-time"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageOrderDetailVO": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetailVO"
            }
          }
        }
      },
...
    }
  }
}